refactor validate and parse
