+++ azurerm/internal/services/web/resource_arm_app_service_virtual_network_swift_connection.go
@@ -55,7 +55,7 @@ func resourceArmAppServiceVirtualNetworkSwiftConnectionCreateUpdate(d *schema.Re
 	ctx, cancel := timeouts.ForCreate(meta.(*clients.Client).StopContext, d)
 	defer cancel()
 
-	appID, err := ParseAppServiceID(d.Get("app_service_id").(string))
+	appID, err := parse.AppServiceID(d.Get("app_service_id").(string))
 	if err != nil {
 		return fmt.Errorf("Error parsing App Service Resource ID %q", appID)
 	}
