+++ azurerm/internal/services/web/validate/app_service.go
@@ -3,6 +3,8 @@ package validate
 import (
 	"fmt"
 	"regexp"
+
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/web/parse"
 )
 
 func AppServiceName(v interface{}, k string) (warnings []string, errors []error) {
@@ -10,6 +12,22 @@ func AppServiceName(v interface{}, k string) (warnings []string, errors []error)
 
 	if matched := regexp.MustCompile(`^[0-9a-zA-Z-]{1,60}$`).Match([]byte(value)); !matched {
 		errors = append(errors, fmt.Errorf("%q may only contain alphanumeric characters and dashes and up to 60 characters in length", k))
+	}
+
+	return warnings, errors
+}
+
+// AppServiceID validates that the specified ID is a valid App Service ID
+func AppServiceID(i interface{}, k string) (warnings []string, errors []error) {
+	v, ok := i.(string)
+	if !ok {
+		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
+		return
+	}
+
+	if _, err := parse.AppServiceID(v); err != nil {
+		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
+		return
 	}
 
 	return warnings, errors
